--- next.config.mjs
@@ -1,5 +1,0 @@
-const nextConfig = {
-  /* config options here */
-};
-
-export default nextConfig;
